requirements.txt
python-fasthtml==0.6.10
click==8.1.7
